+++ src/aggregator/nn_24.cc
@@ -62,7 +62,7 @@ void nn_24::extractUsefulData(kvDataList & out, const kvDataList & dataIn, const
 		throw std::runtime_error("Unable to find correct periods for agregation");
 }
 
-float nn_24::calculate(const ValueList & source, CalculationDataType calcDataType, ExtraData extraData) const
+double nn_24::calculate(const ValueList & source, CalculationDataType calcDataType, ExtraData extraData) const
 {
 	// All values "9" should be interpreted as "8"
 	ValueList modified = source;
