+++ src/aggregator/rr.cc
@@ -58,10 +58,10 @@ void rr::extractUsefulData(kvDataList & out, const kvDataList & dataIn, const kv
 }
 
 
-float rr::calculate(const ValueList & source, CalculationDataType, ExtraData ) const
+double rr::calculate(const ValueList & source, CalculationDataType, ExtraData ) const
 {
 	bool nothing = true;
-	float sum = 0;
+	double sum = 0;
 	for ( ValueList::const_iterator it = source.begin(); it != source.end(); ++ it )
 	{
 		if (* it >= 0)
