+++ src/aggregator/rr_1.h
@@ -62,7 +62,7 @@ public:
 	virtual bool shouldProcess(const kvalobs::kvData &trigger,
 			const kvDataList &observations) const;
 
-	float calculate(const ValueList & source, CalculationDataType calcDataType, ExtraData extraData) const;
+	double calculate(const ValueList & source, CalculationDataType calcDataType, ExtraData extraData) const;
 
 	virtual void extractUsefulData(kvDataList & out, const kvDataList & dataIn, const kvalobs::kvData & trigger) const;
 };
