+++ src/aggregator/uu_24.cc
@@ -45,9 +45,9 @@ uu_24::~uu_24()
 {
 }
 
-float uu_24::calculateWithKoppensFormula(const ValueList & source, float koppenFactor, CalculationDataType, ExtraData ) const
+double uu_24::calculateWithKoppensFormula(const ValueList & source, double koppenFactor, CalculationDataType, ExtraData ) const
 {
-	float q = (source[0] + source[2]) / 2.0;
+	double q = (source[0] + source[2]) / 2.0;
 	return q + (koppenFactor * (source[1] - q));
 }
